+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityTablePanelTest.java
@@ -28,6 +28,7 @@ import is.codion.framework.domain.entity.Entities;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.OrderBy;
 import is.codion.framework.domain.entity.attribute.Attribute;
+import is.codion.swing.common.ui.component.table.FilterTableCellRenderer;
 import is.codion.swing.common.ui.component.table.FilterTableColumn;
 import is.codion.swing.common.ui.component.table.FilterTableColumnModel;
 import is.codion.swing.framework.model.SwingEntityTableModel;
@@ -198,6 +199,28 @@ public class EntityTablePanelTest {
 		assertEquals(Employee.DEPARTMENT, orderBy.orderByColumns().get(0).column());
 		assertTrue(orderBy.orderByColumns().get(1).ascending());
 		assertEquals(Employee.NAME, orderBy.orderByColumns().get(1).column());
+	}
+
+	@Test
+	void cellRenderers() {
+		SwingEntityTableModel tableModel = new SwingEntityTableModel(Employee.TYPE, CONNECTION_PROVIDER);
+		tableModel.refresh();
+		FilterTableCellRenderer<String> nameRenderer = EntityTableCellRenderer.builder(Employee.NAME, tableModel).build();
+		FilterTableCellRenderer<Double> commissionRenderer = EntityTableCellRenderer.builder(Employee.COMMISSION, tableModel).build();
+		EntityTablePanel tablePanel = new EntityTablePanel(tableModel, config -> config
+						.cellRenderer(Employee.NAME, nameRenderer)
+						.cellRendererFactory(new EntityTableCellRenderer.Factory() {
+							@Override
+							public <T> FilterTableCellRenderer<T> create(Attribute<T> attribute, SwingEntityTableModel tableModel) {
+								if (attribute.equals(Employee.COMMISSION)) {
+									return (FilterTableCellRenderer<T>) commissionRenderer;
+								}
+
+								return EntityTableCellRenderer.builder(attribute, tableModel).build();
+							}
+						}));
+		assertSame(nameRenderer, tablePanel.table().columnModel().column(Employee.NAME).getCellRenderer());
+		assertSame(commissionRenderer, tablePanel.table().columnModel().column(Employee.COMMISSION).getCellRenderer());
 	}
 
 	private static List<Entity> initTestEntities(Entities entities) {