+++ changelog.md
@@ -20,6 +20,8 @@ Codion Change Log
 - EntityComboBox.model() added for consistency.
 - EntityComboBox.createForeignKeyFilterControl() removed.
 - EntityTablePanel, no longer removes tooltips from south toolbar buttons, seem to only block mouse cursor in MetalLookAndFeel.
+- EntityTableCellRenderer bug fixed, create(Attribute, SwingEntityTableModel) never got called.
+- EntityTable.Config.cellEditorFactory() and cellRendererFactory() added.
 
 ## 0.18.18
 ### is.codion.common.db